### 2021-07-25 #3536 @whedon
[PRE REVIEW]: sorn: A Python package for Self Organizing Recurrent Neural Network
**Submitting author:** @saran-nns (<a href="http://orcid.org/0000-0002-7314-0261">Saranraj Nambusubramaniyan</a>)
**Repository:** <a href="https://github.com/Saran-nns/sorn/" target ="_blank">https://github.com/Saran-nns/sorn/</a>
**Version:** 0.4.9
