+++ paper/paper.bib
@@ -16,6 +16,7 @@
   number={1},
   pages={e1002848},
   year={2013},
+  doi = {10.1371/journal.pcbi.1002848},
   publisher={Public Library of Science San Francisco, USA}
 }
 
@@ -27,6 +28,7 @@
   number={5},
   pages={1--21},
   year={2017},
+  doi = {10.1371/journal.pone.0178683},
   publisher={Public Library of Science}
 }
 
