### 2021-07-26 881b22a
fixing typo in paper.md
