### 2021-07-26 #3536 @rkurchin
Thanks! @KirstensGitHub and @agricolab, would you be willing to review this submission to JOSS?

We carry out our checklist-driven reviews here in GitHub issues according to these guidelines. You can find out more about this submission in the first comment in this issue.
