### 2021-07-27 #3536 @rkurchin
Hi @AlexChristensen, would you be willing to review this submission to JOSS?

JOSS publishes articles about open source research software. It is a free, open-source, community driven and developer-friendly online journal (no publisher is seeking to raise revenue from the volunteer labor of researchers!). The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html. Editors and reviewers post comments on the Review issue (this is the "pre-review" issue), and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review).
