### 2021-07-28 #3536 @AlexChristensen
Hi @rkurchin,

Although I am familiar with neural networks, I do not feel I have the expertise to sufficiently evaluate the Python code. My experience is in the application of Python functions, but I have little experience writing Python code.

I appreciate the invitation and look forward to seeing @Saran-nns's result.
