### 2021-07-29 #3545 @whedon
[REVIEW]: sorn: A Python package for Self Organizing Recurrent Neural Network
**Submitting author:** @saran-nns (<a href="http://orcid.org/0000-0002-7314-0261">Saranraj Nambusubramaniyan</a>)
**Repository:** <a href="https://github.com/Saran-nns/sorn/" target ="_blank">https://github.com/Saran-nns/sorn/</a>
**Version:** v0.6.2-beta

- [x] #3547

- [x] #3554
