### 2021-07-30 Saran-nns/sorn#26 @janfb
[JOSS review] is the list of authors in the JOSS submission complete?
Hi there, another question related to the review (https://github.com/openjournals/joss-reviews/issues/3545).

this code is the basis for your master thesis research conducted at University of Osnabrück, correct? Then I was wondering whether there could be any other researches involved in the work underlying this package, e.g., thesis supervisors?
