### 2021-07-30 b981078
travis update: multiple OS, py vers test
