+++ .travis.yml
@@ -1,23 +1,31 @@
-language:
-  python
+language: python
+
+os:
+  - linux
+  - osx
 
 python:
   - "3.5"
+  - "3.6"
+  - "3.7"
+  - "3.8"
+  - "3.9"
 
-cache: 
-  pip
+cache: pip
 
 install:
   - pip install -r requirements.txt
 
 before_script:
-- pip install coverage
-- pip install coveralls
-- pip install codecov
+  - python --version
+  - python install -U pip
+  - pip install coverage
+  - pip install coveralls
+  - pip install codecov
 
 script:
   - python test_sorn.py
-  - coverage run test_sorn.py  
+  - coverage run test_sorn.py
 
 after_success:
-- codecov
+  - codecov
