+++ .travis.yml
@@ -1,15 +1,39 @@
 language: python
 
-os:
-  - linux
-  - osx
+jobs:
+  include:
+    - os: linux
+      python: 3.5
+      env: TOXENV=py35
+    - os: linux
+      python: 3.6
+      env: TOXENV=py36
+    - os: linux
+      python: 3.7
+      env: TOXENV=py37
+    - os: linux
+      python: 3.8
+      env: TOXENV=py38
 
-python:
-  - "3.5"
-  - "3.6"
-  - "3.7"
-  - "3.8"
-  - "3.9"
+    - os: linux
+      python: 3.9
+      env: TOXENV=py39
+
+    - os: osx
+      language: generic
+      env: TOXENV=py35
+    - os: osx
+      language: generic
+      env: TOXENV=py36
+    - os: osx
+      language: generic
+      env: TOXENV=py37
+    - os: osx
+      language: generic
+      env: TOXENV=py38
+    - os: osx
+      language: generic
+      env: TOXENV=py39
 
 cache: pip
 
@@ -18,7 +42,6 @@ install:
 
 before_script:
   - python --version
-  - python install -U pip
   - pip install coverage
   - pip install coveralls
   - pip install codecov
