### 2021-07-30 fa9afdf
Update sorn.py
