### 2021-07-30 a529c33
Update sorn.py
