### 2021-07-30 97b3269
update numpy, pandas, matplotlib, seaborn versions
