+++ requirements.txt
@@ -1,14 +1,10 @@
-certifi==2018.11.29
-cycler==0.10.0
-kiwisolver==1.0.1
-matplotlib==3.0.3
-numpy==1.16.2
-pandas==0.24.1
+numpy==1.19.5
+pandas==1.1.5
 pyparsing==2.3.1
 python-dateutil==2.8.0
 pytz==2018.9
-scipy==1.2.1
-seaborn==0.9.0
+scipy==1.4.1
+seaborn==0.11.1
 six==1.12.0
 wincertstore==0.2
-networkx==2.4
+networkx==2.5.1
