+++ .travis.yml
@@ -15,10 +15,6 @@ jobs:
       python: 3.8
       env: TOXENV=py38
 
-    - os: linux
-      python: 3.9
-      env: TOXENV=py39
-
     - os: osx
       language: generic
       env: TOXENV=py35
@@ -31,9 +27,6 @@ jobs:
     - os: osx
       language: generic
       env: TOXENV=py38
-    - os: osx
-      language: generic
-      env: TOXENV=py39
 
 cache: pip
 
