+++ requirements.txt
@@ -1,10 +1,10 @@
-numpy==1.19.5
-pandas==1.1.5
+numpy==1.18.5
+pandas==1.2.0
 pyparsing==2.3.1
 python-dateutil==2.8.0
 pytz==2018.9
-scipy==1.4.1
-seaborn==0.11.1
+scipy==1.4.0
+seaborn==0.11.0
 six==1.12.0
 wincertstore==0.2
-networkx==2.5.1
+networkx==2.4.1
