+++ requirements.txt
@@ -1,5 +1,5 @@
 numpy==1.18.5
-pandas==1.2.0
+pandas==0.25.3
 pyparsing==2.3.1
 python-dateutil==2.8.0
 pytz==2018.9
