+++ requirements.txt
@@ -4,7 +4,7 @@ pyparsing==2.3.1
 python-dateutil==2.8.0
 pytz==2018.9
 scipy==1.4.0
-seaborn==0.11.0
+seaborn==0.9.1
 six==1.12.0
 wincertstore==0.2
 networkx==2.4.1
