+++ requirements.txt
@@ -6,5 +6,5 @@ pytz==2018.9
 scipy==1.4.0
 seaborn==0.9.1
 six==1.12.0
-wincertstore==2.3
+wincertstore==0.2
 networkx==2.4.1
