+++ requirements.txt
@@ -7,4 +7,4 @@ scipy==1.4.0
 seaborn==0.9.1
 six==1.12.0
 wincertstore==0.2
-networkx==2.4.1
+networkx==2.4
