+++ .travis.yml
@@ -14,6 +14,9 @@ jobs:
     - os: linux
       python: 3.8
       env: TOXENV=py38
+    - os: linux
+      python: 3.9
+      env: TOXENV=py39
 
     - os: osx
       language: generic
