+++ .travis.yml
@@ -14,22 +14,23 @@ jobs:
     - os: linux
       python: 3.8
       env: TOXENV=py38
-    - os: linux
-      python: 3.9
-      env: TOXENV=py39
 
-    - os: osx
-      language: generic
-      env: TOXENV=py35
-    - os: osx
-      language: generic
-      env: TOXENV=py36
-    - os: osx
-      language: generic
-      env: TOXENV=py37
-    - os: osx
-      language: generic
-      env: TOXENV=py38
+    - name: "Python 3.7.5 on macOS 10.14"
+      os: osx
+      osx_image: xcode10.2
+      language: shell
+      before_script:
+        - python --version
+        - pip install coverage
+        - pip install coveralls
+        - pip install codecov
+
+      script:
+        - python test_sorn.py
+        - coverage run test_sorn.py
+
+      after_success:
+        - codecov
 
 cache: pip
 
