+++ .travis.yml
@@ -15,12 +15,29 @@ jobs:
       python: 3.8
       env: TOXENV=py38
 
+    - name: "Python 3.6.5 on macOS 10.13"
+      os: osx
+      osx_image: xcode9.4
+      language: shell
+      before_script:
+        - python3 --version
+        - pip install coverage
+        - pip install coveralls
+        - pip install codecov
+
+      script:
+        - python test_sorn.py
+        - coverage run test_sorn.py
+
+      after_success:
+        - codecov
+
     - name: "Python 3.7.5 on macOS 10.14"
       os: osx
       osx_image: xcode10.2
       language: shell
       before_script:
-        - python --version
+        - python3 --version
         - pip install coverage
         - pip install coveralls
         - pip install codecov
