+++ .travis.yml
@@ -18,6 +18,8 @@ jobs:
     - name: "Python 3.6.5 on macOS 10.13"
       os: osx
       osx_image: xcode10.2
+      dist: xenial
+      language: shell
       before_script:
         - python3 --version
         - python3 install coverage
@@ -34,6 +36,7 @@ jobs:
     - name: "Python 3.7.5 on macOS 10.14"
       os: osx
       osx_image: xcode11.2
+      dist: xenial
       language: shell
       before_script:
         - python3 --version
