### 2021-07-31 Saran-nns/sorn#26 @Saran-nns
Hi @janfb. The network was created as part of my master's thesis. The actual thesis can be found at [repo](https://github.com/Saran-nns/PySORN_0.1). However, the PyPI package is not part of my thesis. Also, I have informed my supervisor about a possible publication in the future, but have not received any feedback in the last years. However, I am happy to credit my supervisor as a co-author of this package. But I think it is not appropriate to act without his concern. I will follow your recommendation in this case.
