### 2021-07-31 Saran-nns/sorn#25 @Saran-nns
Hi @janfb I checked the problem with the Windows and Ubuntu machines, the code works without problems up to py3.8 and for py3.9 there are version compatibility issues between dependencies.
I have also checked the Mac OS virtual environments and there are no problems when the package is imported at Jupyter notebooks.
However, when I import the package from the terminal, I got errors and was able to reproduce the same as you reported. I suspect that the problem might be related to the system 'PATH'. Thank you for reporting the problem. I am investigating it.
