+++ build/lib/sorn/__init__.py
@@ -1,8 +1,0 @@
-from .sorn import Simulator, Trainer
-import logging
-from .utils import *
-
-__author__ = "Saranraj Nambusubramaniyan"
-__version__ = "0.4.9"
-
-logging.basicConfig(level=logging.INFO)
