+++ sorn.egg-info/requires.txt
@@ -1,4 +1,0 @@
-numpy
-configparser
-scipy
-seaborn
