+++ build/lib/sorn/sorn.py
@@ -1,1209 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import division
-import numpy as np
-import os
-from configparser import ConfigParser
-import random
-
-try:
-    from sorn.utils import Initializer
-except:
-    from utils import Initializer
-
-parser = ConfigParser()
-config_file = os.path.join(
-    os.path.dirname(os.path.abspath(__file__)), "configuration.ini"
-)
-parser.read(config_file)
-
-
-class Sorn(object):
-
-    """ This class wraps initialization of the network and its parameters
-
-    Args:
-        nu (int): Number of input units. Defaults to 10
-
-        ne (int): Sorn._ne  # Number of excitatory units. Defaults to 200
-
-        eta_stdp (float): STDP plasticity Learning rate constant; SORN1 and SORN2. Defaults to 0.004
-
-        eta_ip (float): Intrinsic plasticity learning rate constant; SORN1 and SORN2. Defaults to 0.001
-
-        eta_inhib (float): Intrinsic plasticity learning rate constant; SORN2 only. Defaults to 0.01
-
-        h_ip (float): Target firing rate. Defaults to 2 * Sorn._nu / Sorn._ne
-
-        mu_ip (float): Mean target firing rate. Defaults to 0.1
-
-        sigma_ip (float): Variance of target firing rate. Defaults to 0.0
-
-        ni (int): Number of inhibitory units in the network. Defaults to int(0.2 * Sorn._ne)
-
-        time_steps (float): Total time steps of simulation
-
-        te_min (float): Excitatory minimum Threshold. Defaults to 0.0
-
-        te_max (float): Excitatory maximum Threshold. Defaults to 1.0
-
-        ti_min (float): Inhibitory minimum Threshold. Defaults to 0.0
-
-        ti_max (float): Inhibitory maximum Threshold. Defaults to 0.5
-
-        network_type_ee (str): Dense or Sparse. Defaults to Sparse
-
-        network_type_ei (str): Dense or Sparse. Defaults to Sparse
-
-        network_type_ie (str): Dense or Sparse. Defaults to Dense
-
-        lambda_ee (int):  Number of connections to and from a single excitatory unit to another at initialization. Defaults to 20
-
-        lambda_ei (int): Number of connections to and from a single inhibitory unit to exitatory unit at initialization. Defaults to 40
-
-        lambda_ie (int):  Number of connections to and from a single excitatory unit to inhibitory unit at initialization. Defaults to 100"""
-
-    def __init__(self):
-        pass
-
-    _nu = int(parser.get("Network_Config", "Nu"))
-    _ne = int(parser.get("Network_Config", "Ne"))
-    _ni = int(0.2 * _ne)
-    _eta_stdp = float(parser.get("Network_Config", "eta_stdp"))
-    _eta_inhib = float(parser.get("Network_Config", "eta_inhib"))
-    _eta_ip = float(parser.get("Network_Config", "eta_ip"))
-    _te_max = float(parser.get("Network_Config", "te_max"))
-    _ti_max = float(parser.get("Network_Config", "ti_max"))
-    _ti_min = float(parser.get("Network_Config", "ti_min"))
-    _te_min = float(parser.get("Network_Config", "te_min"))
-    _mu_ip = float(parser.get("Network_Config", "mu_ip"))
-    _sigma_ip = float(parser.get("Network_Config", "sigma_ip"))
-    _network_type_ee = str(parser.get("Network_Config", "network_type_ee"))
-    _network_type_ei = str(parser.get("Network_Config", "network_type_ei"))
-    _network_type_ie = str(parser.get("Network_Config", "network_type_ie"))
[…]
