+++ build/lib/sorn/utils.py
@@ -1,1201 +1,0 @@
-from __future__ import division
-import numpy as np
-from scipy.stats import norm
-import random
-import matplotlib.pyplot as plt
-import seaborn as sns
-from scipy.optimize import curve_fit
-from scipy import stats
-import networkx as nx
-import pandas as pd
-from mpl_toolkits.axes_grid1.inset_locator import InsetPosition
-
-
-class Initializer(object):
-    """
-    Helper class to initialize the matrices for the SORN
-    """
-
-    def __init__(self):
-        pass
-
-    @staticmethod
-    def generate_strong_inp(length: int, reservoir_size: int):
-        """Generate strong one-hot vector of input. Random neurons in the reservoir acts as inputs
-
-        Args:
-            length (int): Number of input neurons
-
-        Returns:
-            inp (array): Input vector of length equals the number of neurons in the reservoir
-                  with randomly chosen neuron set active
-
-            idx (list): List of chosen input neurons """
-
-        inp = [0] * reservoir_size
-        x = [0] * length
-        idx = np.random.choice(length, np.random.randint(reservoir_size))
-
-        for i in idx:
-            x[i] = 1.0e4
-
-        inp[: len(x)] = x
-
-        return inp, idx
-
-    # Generate multi-node one-hot strong inputs
-
-    @staticmethod
-    def multi_one_hot_inp(ne: int, inputs: list, n_nodes_per_inp: int):
-        """Generate multi(n_nodes_per_inp) one hot vector for each input.
-        For each input, set n_nodes_per_inp equals one and the rest of
-        neurons in the pool recieves no external stimuli
-
-        Args:
-          ne (int): Number of excitatory units in sorn
-
-          inputs (list): input labels
-
-          n_nodes_per_inp(int): Number of target units in pool that receives single input
-
-        Returns:
-          one_hot_vector for each label with length equals ne
-
-        """
-
-        one_hot = np.zeros((ne, len(inputs)))
-
-        idxs = []
-
-        for _ in range(n_nodes_per_inp):
-            idxs.append(random.sample(range(0, ne), len(inputs)))
-
-        idxs = list(zip(*idxs))
-
-        j = 0  # Max(j) = len(inputs)
-        for idx_list in idxs:
-            for i in idx_list:
-                one_hot[i][j] = 1
-            j += 1
-
-        return one_hot, idxs
-
-    @staticmethod
-    def generate_gaussian_inputs(length: int, reservoir_size: int):
-
-        """Generate external stimuli sampled from Gaussian distribution.
-        Randomly neurons in the reservoir receives this input at each timestep
-
-        Args:
-            length (int): Number of input neurons
-
-        Returns:
-            out (array): Input vector of length equals the number of neurons in the reservoir
-                  with randomly chosen neuron set active
-
-            idx (int): List of chosen input neurons
-        """
-
-        out = [0] * reservoir_size
-        x = [0] * length
-        idx = np.random.choice(length, np.random.randint(reservoir_size))
-        inp = np.random.normal(length)
-
-        for i in idx:
-            x[i] = inp[i]
-
-        out[: len(x)] = x
-
-        return out, idx
-
-    @staticmethod
-    def normalize_weight_matrix(weight_matrix: np.array):
-
-        # Applied only while initializing the weight. During simulation, Synaptic scaling applied on weight matrices
-
-        """ Normalize the weights in the matrix such that incoming connections to a neuron sum up to 1
-
-        Args:
-            weight_matrix (array): Incoming Weights from W_ee or W_ei or W_ie
-
-        Returns:
-            weight_matrix (array): Normalized weight matrix"""
-
-        normalized_weight_matrix = weight_matrix / np.sum(weight_matrix, axis=0)
-
-        return normalized_weight_matrix
-
-    @staticmethod
[…]
