+++ requirements.txt
@@ -2,8 +2,8 @@ certifi==2018.11.29
 cycler==0.10.0
 kiwisolver==1.0.1
 matplotlib==3.0.3
-numpy==1.16.2
-pandas==0.24.1
+numpy==1.19.2
+pandas==1.3.1
 pyparsing==2.3.1
 python-dateutil==2.8.0
 pytz==2018.9
