+++ build/lib/sorn/__init__.py
@@ -1,0 +1,8 @@
+from .sorn import Simulator, Trainer
+import logging
+from .utils import *
+
+__author__ = "Saranraj Nambusubramaniyan"
+__version__ = "0.5.1"
+
+logging.basicConfig(level=logging.INFO)
