+++ sorn.egg-info/SOURCES.txt
@@ -1,0 +1,14 @@
+LICENSE.md
+README.md
+setup.py
+sorn/__init__.py
+sorn/configuration.ini
+sorn/sorn.py
+sorn/test_sorn.py
+sorn/utils.py
+sorn.egg-info/PKG-INFO
+sorn.egg-info/SOURCES.txt
+sorn.egg-info/dependency_links.txt
+sorn.egg-info/not-zip-safe
+sorn.egg-info/requires.txt
+sorn.egg-info/top_level.txt
