+++ build/lib/sorn/configuration.ini
@@ -1,0 +1,38 @@
+[Network_Config]
+
+# Number of neurons in the reservoir
+Ne = 200
+
+# Number of input units in the reservoir
+Nu = 10
+
+# Connection/Network type: Sparse or Dense
+
+network_type_ee = Sparse
+network_type_ei = Sparse
+network_type_ie = Dense
+
+# Mean number of incoming and outgoing connections per neuron
+lambda_ee = 20
+lambda_ei = 40
+lambda_ie = 100
+
+# Hebbian Learning rate 
+eta_stdp = 0.004
+
+# Anti-Hebbian learning rate
+eta_inhib = 0.001
+
+# Intrinsic plasticity constants    
+eta_ip = 0.01
+
+# Threshold initialization values
+te_max = 1.0
+ti_max = 0.5
+ti_min = 0.0
+te_min = 0.0
+
+# Mean and standard deviation of target firing rate , h_ip
+mu_ip = 0.1
+sigma_ip = 0.0 
+
