+++ requirements.txt
@@ -1,14 +1,16 @@
-certifi==2018.11.29
+certifi==2021.5.30
+configparser==5.0.2
 cycler==0.10.0
-kiwisolver==1.0.1
-matplotlib==3.0.3
-numpy==1.19.2
+networkx==2.6.2
+kiwisolver==1.3.1
+matplotlib==3.4.2
+numpy==1.21.1
 pandas==1.3.1
-pyparsing==2.3.1
-python-dateutil==2.8.0
-pytz==2018.9
-scipy==1.2.1
-seaborn==0.9.0
-six==1.12.0
+pyparsing==2.4.7
+python-dateutil==2.8.2
+pytz==2021.1
+scipy==1.7.1
+seaborn==0.11.1
+six==1.16.0
 wincertstore==0.2
-networkx==2.4
+
