### 2021-08-02 Saran-nns/sorn#25 @Saran-nns
Hi @janfb, `networkx` package is added to the requirements. I made the necessary changes and updated the pypi package v0.5.1 is online now. Thanks again for the issue.
