### 2021-08-05 Saran-nns/sorn#26 @janfb
Hi @Saran-nns okay, I think this is fine then. I just wanted to make sure the supervisor(s) were considered in this process. Thanks!
