### 2021-08-06 Saran-nns/sorn#29 @janfreyberg
[JOSS review] Examples in README fail
Hi there,

I was just trying to run the examples as provided in the README, but ran into an error:

```python
from sorn import Simulator
import numpy as np

# Sample input
num_features = 10
time_steps = 200
inputs = np.random.rand(num_features,time_steps)

# To simulate the network;
matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=None, noise = True, time_steps=time_steps, _ne = 200, _nu=num_features)

# To resume the simulation, load the matrices_dict from previous simulation;
matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=matrices_dict, noise= True, time_steps=time_steps,_ne = 200, _nu=num_features)
```

Produces the following error:
```
Network Initialized
Number of connections in Wee 3881 , Wei 1584, Wie 8000
Shapes Wee (200, 200) Wei (40, 200) Wie (200, 40)
<ipython-input-1-3e49ac279123> in <module>
      8
      9 # To simulate the network;
---> 10 matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=None, noise = True, time_steps=time_steps, _ne = 200, _nu=num_features)
     11
     12 # To resume the simulation, load the matrices_dict from previous simulation;

~/github/joss-reviews/.venv/lib/python3.9/site-packages/sorn/sorn.py in simulate_sorn(self, inputs, phase, matrices, time_steps, noise, freeze, **kwargs)
    956
    957             # STDP
--> 958             if 'stdp' not in self.freeze:
    959                 Wee_t = plasticity.stdp(
    960                     Wee[i], x_buffer, cutoff_weights=(0.0, 1.0))

TypeError: argument of type 'NoneType' is not iterable
```

I assume this is a relatively easy fix for the `freeze` attribute?

Thanks

_this is part of the review: https://github.com/openjournals/joss-reviews/issues/3545_
