### 2021-08-06 Saran-nns/sorn#29 @janfreyberg
Note that when I try to provide an argument for `freeze`, I also encounter errors:

```python
matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(
    inputs = inputs, phase='plasticity', matrices=None, noise = True,
    time_steps=time_steps, _ne = 200, _nu=num_features,
    freeze=['ip'])

Network Initialized
Number of connections in Wee 3904 , Wei 1569, Wie 8000
Shapes Wee (200, 200) Wei (40, 200) Wie (200, 40)
UnboundLocalError                         Traceback (most recent call last)
<ipython-input-4-8c54dfffe88b> in <module>
      8
      9 # To simulate the network;
---> 10 matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=None, noise = True, time_steps=time_steps, _ne = 200, _nu=num_features, freeze=['ip'])

~/github/joss-reviews/.venv/lib/python3.9/site-packages/sorn/sorn.py in simulate_sorn(self, inputs, phase, matrices, time_steps, noise, freeze, **kwargs)
    981             # Assign the matrices to the matrix collections
    982             matrix_collection.weight_matrix(Wee_t, Wei_t, Wie[i], i)
--> 983             matrix_collection.threshold_matrix(Te_t, Ti[i], i)
    984             matrix_collection.network_activity_t(x_buffer, y_buffer, i)
    985

UnboundLocalError: local variable 'Te_t' referenced before assignment
```
