### 2021-08-06 Saran-nns/sorn#29 @janfreyberg
I should note that I am currently using the latest install from github, but the first error also occurred on the version in PyPi!
