### 2021-08-06 38cedb1
#29 freeze arg check
