### 2021-08-06 Saran-nns/sorn#29 @Saran-nns
Hi @janfreyberg. Thank you for reporting the issue. The variable assignment and the freeze options are fixed. The github repo and the pypi (v0.5.2) are also updated. Test runs of README examples are done at [colab notebook](https://colab.research.google.com/drive/1qz_Xkfam0wCCu2pz2kFogEJDWU2ED8Q4#scrollTo=xpHqf0bijR_d)
 Thank you :)
