### 2021-08-13 Saran-nns/sorn#29 @janfreyberg
Thanks @Saran-nns (and sorry for the delay) - I can run the examples from that notebook, but the following from the README seems to still fail:

```python
matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=None, noise = True, time_steps=time_steps, _ne = 200, _nu=num_features, freeze=['ip'])
```

Seems like the main difference between this and your colab is the `time_steps` argument, which is set to `time_steps` in the README, but 1 in the colab.

Can you clarify whether this should work as is in the readme and fix a bug if so - or change the readme otherwise?

You might also want to use some way of testing your README examples automatically. This library could help with that:
https://github.com/koaning/mktestdocs
