+++ sorn/sorn.py
@@ -3,20 +3,12 @@
 from __future__ import division
 import numpy as np
 import os
-from configparser import ConfigParser
 import random
 
 try:
     from sorn.utils import Initializer
 except:
     from utils import Initializer
-
-parser = ConfigParser()
-config_file = os.path.join(
-    os.path.dirname(os.path.abspath(__file__)), "configuration.ini"
-)
-parser.read(config_file)
-
 
 class Sorn(object):
 
@@ -64,29 +56,27 @@ class Sorn(object):
         lambda_ie (int):  Number of connections to and from a single excitatory unit to inhibitory unit at initialization. Defaults to 100"""
 
     def __init__(self):
-        pass
 
-    _nu = int(parser.get("Network_Config", "Nu"))
-    _ne = int(parser.get("Network_Config", "Ne"))
-    _ni = int(0.2 * _ne)
-    _eta_stdp = float(parser.get("Network_Config", "eta_stdp"))
-    _eta_inhib = float(parser.get("Network_Config", "eta_inhib"))
-    _eta_ip = float(parser.get("Network_Config", "eta_ip"))
-    _te_max = float(parser.get("Network_Config", "te_max"))
-    _ti_max = float(parser.get("Network_Config", "ti_max"))
-    _ti_min = float(parser.get("Network_Config", "ti_min"))
-    _te_min = float(parser.get("Network_Config", "te_min"))
-    _mu_ip = float(parser.get("Network_Config", "mu_ip"))
-    _sigma_ip = float(parser.get("Network_Config", "sigma_ip"))
-    _network_type_ee = str(parser.get("Network_Config", "network_type_ee"))
-    _network_type_ei = str(parser.get("Network_Config", "network_type_ei"))
-    _network_type_ie = str(parser.get("Network_Config", "network_type_ie"))
-    _lambda_ee = int(parser.get("Network_Config", "lambda_ee"))
-    _lambda_ei = int(parser.get("Network_Config", "lambda_ei"))
-    _lambda_ie = int(parser.get("Network_Config", "lambda_ie"))
+        self.nu = 10
+        self.ne = 200
+        self.ni = int(0.2 * self.ne)
+        self.eta_stdp = 0.004
+        self.eta_inhib = 0.001
+        self.eta_ip = 0.1
+        self.te_max = 1.0
+        self.ti_max = 0.5
+        self.ti_min = 0.0
+        self.te_min = 0.0
+        self.mu_ip = 0.1
+        self.sigma_ip = 0.0
+        self.network_type_ee = "Sparse"
+        self.network_type_ei = "Sparse"
+        self.network_type_ie = "Dense"
+        self.lambda_ee = 20
+        self.lambda_ei = 40
+        self.lambda_ie = 100
 
-    @staticmethod
-    def initialize_weight_matrix(
+    def initialize_weight_matrix(self,
         network_type: str, synaptic_connection: str, self_connection: str, lambd_w: int
     ):
         """Wrapper for initializing the weight matrices for SORN
@@ -104,22 +94,21 @@ class Sorn(object):
         if (network_type == "Sparse") and (self_connection == "False"):
 
             # Generate weight matrix for E-E/ E-I connections with mean lamda incoming and out-going connections per neuron
-            assert (lambd_w <= Sorn._ne), "Number of connections per unit (lambda) should be less than number of units(Ne) in the pool and also Ne should be greater than 25"
+            assert (lambd_w <= self.ne), "Number of connections per unit (lambda) should be less than number of units(Ne) in the pool and also Ne should be greater than 25"
             weight_matrix = Initializer.generate_lambd_connections(
-                synaptic_connection, Sorn._ne, Sorn._ni, lambd_w, lambd_std=1
[…]
