+++ sorn/__init__.py
@@ -3,6 +3,6 @@ import logging
 from .utils import *
 
 __author__ = "Saranraj Nambusubramaniyan"
-__version__ = "0.5.2"
+__version__ = "0.5.3"
 
 logging.basicConfig(level=logging.INFO)
