+++ .readthedocs.yml
@@ -13,8 +13,8 @@ sphinx:
 formats: all
 
 python:
-  version: 3.6
+  version: 3.8
   install:
-     - requirements: docs/requirements.txt
-     - method: pip
-       path: .
+    - requirements: docs/requirements.txt
+    - method: pip
+      path: .
