+++ .pre-commit-config.yaml
@@ -1,11 +1,11 @@
 repos:
-- repo: https://github.com/ambv/black
-  rev: stable
-  hooks:
-    - id: black
-      language_version: python3.5
-- repo: https://gitlab.com/pycqa/flake8
-  rev: 3.8.3
-  hooks: 
-    - id: flake8
-      language: python_venv
+  - repo: https://github.com/ambv/black
+    rev: stable
+    hooks:
+      - id: black
+        language_version: python3.8
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: 3.8.3
+    hooks:
+      - id: flake8
+        language: python_venv
