### 2021-08-14 fc64266
issue #31 config, cartpole, readme and figures update
