+++ sorn/utils.py
@@ -424,7 +424,7 @@ class Plotter(object):
     def hist_incoming_conn(
         weights: np.array, bin_size: int, histtype: str, savefig: bool
     ):
-        """Plot the histogram of number of incoming connections per neuron
+        """Plot the histogram of number of presynaptic connections per neuron
 
         Args:
             weights (array): Connection weights
@@ -438,35 +438,29 @@ class Plotter(object):
         Returns:
             plot (matplotlib.pyplot): plot object
         """
-
-        # Plot the histogram of distribution of number of incoming connections in the network
-
         num_incoming_weights = np.sum(np.array(weights) > 0, axis=0)
 
         plt.figure(figsize=(12, 5))
-
-        plt.title("Number of incoming connections")
         plt.xlabel("Number of connections")
-        plt.ylabel("Count")
-        plt.hist(num_incoming_weights, bins=bin_size, histtype=histtype)
+        plt.ylabel("Probability")
 
-        # Empirical average and variance are computed
-        avg = np.mean(num_incoming_weights)
-        var = np.var(num_incoming_weights)
-        # From hist plot above, it is clear that connection count follow gaussian distribution
-        pdf_x = np.linspace(
-            np.min(num_incoming_weights), np.max(num_incoming_weights), 100
-        )
-        pdf_y = 1.0 / np.sqrt(2 * np.pi * var) * np.exp(-0.5 * (pdf_x - avg) ** 2 / var)
+        # Fit a normal distribution to the data
+        mu, std = norm.fit(num_incoming_weights)
+        plt.hist(num_incoming_weights, bins=bin_size, density=True, alpha=0.6, color='b')
 
-        plt.plot(pdf_x, pdf_y, "k--", label="Gaussian fit")
-        plt.axvline(x=avg, color="r", linestyle="--", label="Mean")
-        plt.legend()
+        # PDF
+        xmin, xmax = plt.xlim()
+        x = np.linspace(xmin, xmax, max(num_incoming_weights))
+        p = norm.pdf(x, mu, std)
+        plt.plot(x, p, 'k', linewidth=2)
+        title = "Distribution of presynaptic connections: mu = %.2f,  std = %.2f" % (mu, std)
+        plt.title(title)
 
         if savefig:
             plt.savefig("hist_incoming_conn")
 
         return plt.show()
+
 
     @staticmethod
     def hist_outgoing_conn(
@@ -491,23 +485,20 @@ class Plotter(object):
         num_outgoing_weights = np.sum(np.array(weights) > 0, axis=1)
 
         plt.figure(figsize=(12, 5))
-
-        plt.hist(num_outgoing_weights, bins=bin_size, histtype=histtype)
-        plt.title("Number of Outgoing connections")
         plt.xlabel("Number of connections")
-        plt.ylabel("Count")
-        # Empirical average and variance are computed
-        avg = np.mean(num_outgoing_weights)
-        var = np.var(num_outgoing_weights)
+        plt.ylabel("Probability")
 
-        pdf_x = np.linspace(
-            np.min(num_outgoing_weights), np.max(num_outgoing_weights), 100
-        )
-        pdf_y = 1.0 / np.sqrt(2 * np.pi * var) * np.exp(-0.5 * (pdf_x - avg) ** 2 / var)
+        # Fit a normal distribution to the data
+        mu, std = norm.fit(num_outgoing_weights)
+        plt.hist(num_outgoing_weights, bins=bin_size, density=True, alpha=0.6, color='b')
 
-        plt.plot(pdf_x, pdf_y, "k--", label="Gaussian fit")
-        plt.axvline(x=avg, color="r", linestyle="--", label="Mean")
-        plt.legend()
+        # PDF
+        xmin, xmax = plt.xlim()
[…]
