+++ README.md
@@ -61,31 +61,11 @@ To install all optional dependencies,
 For detailed documentation about usage and development, please visit [SORN-Documentation](https://self-organizing-recurrent-neural-networks.readthedocs.io/)
 
 ## Usage
-
-### Update Network configurations
-There are two ways to update/configure the network parameters,
-1. Navigate to home/conda/envs/ENVNAME/Lib/site-packages/sorn
- ```or``` if you are unsure about the directory of ```sorn```
-
-Run
+### Plasticity Phase
+The default ```ne, nu``` values are overriden by passing them as kwargs inside ```simulate_sorn``` method.
 
 ```python
 import sorn
-
-sorn.__file__
-```
-to find the location of the sorn package
-
-Then, update/edit arguments in `configuration.ini`
-
-2. While instantiating the network using `Simulator` or `Trainer` objects, override the defaults by assigning the values to the ```kwargs``` shown below,
-```Python
-kwargs = {'_ne', '_nu', '_network_type_ee', '_network_type_ei', '_network_type_ie', '_lambda_ee','_lambda_ei', '_lambda_ie', '_eta_stdp','_eta_inhib', '_eta_ip', '_te_max', '_ti_max', '_ti_min', '_te_min', '_mu_ip','_sigma_ip'}
-```
-### Plasticity Phase
-The default ```_ne, _nu``` values are overriden by passing them as kwargs inside ```simulate_sorn``` method.
-
-```Python
 from sorn import Simulator
 import numpy as np
 
@@ -94,11 +74,42 @@ num_features = 10
 time_steps = 200
 inputs = np.random.rand(num_features,time_steps)
 
-# To simulate the network;
-matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=None, noise = True, time_steps=time_steps, _ne = 200, _nu=num_features)
+# Simulate the network with default hyperparameters under gaussian white noise
+state_dict, E, I, R, C = Simulator.simulate_sorn(inputs = inputs, phase='plasticity',
+                                                matrices=None, noise = True,
+                                                time_steps=time_steps)
 
-# To resume the simulation, load the matrices_dict from previous simulation;
-matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=matrices_dict, noise= True, time_steps=time_steps,_ne = 200, _nu=num_features)
+```
+The network defaults are,
+
+Keyword argument | Description | Default Value |
+--- | --- | --- |
+ne              | 200      | Number of Encitatory neurons in the reservoir                              |
+nu              | 10       | Number of Input neurons in the reservoir                                   |
+network_type_ee | "Sparse" | `Sparse` or `Dense` connectivity between Excitatory neurons                |
+network_type_ie | "Dense"  | `Sparse` or `Dense` connectivity from Excitatory to Inhibitory neurons     |
+network_type_ei | "Sparse" | `Sparse` or `Dense` connectivity from Inhibitory to Excitatory neurons     |
+lambda_ee       | 20       | % of connections between neurons in Excitatory pool                        |
+lambda_ei       | 40       | % of connections from Inhibitory to Excitatory neurons                     |
+lambda_ie       | 100      | % of connections from Excitatory to Inhibitory neurons                     |
+eta_stdp        | 0.004    | Hebbian Learning rate for connections between excitatory neurons           |
+eta_inhib       | 0.001    | Hebbian Learning rate for connections from Inhibitory to Excitatory neurons|
[…]
