+++ docs/source/usage.rst
@@ -1,123 +1,381 @@
 Usage
 =====
 
-Update network configuration
-----------------------------
-
-Configuration file
-^^^^^^^^^^^^^^^^^^
-Navigate to `home/conda/envs/ENVNAME/Lib/site-packages/sorn` or if you are unsure about the directory of `sorn`:
-
-Run::
+Plasticity Phase
+-----------------
 
     import sorn
-    sorn.__file__
-
-to find the location of the sorn package:
-
-Then, update/edit arguments in `configuration.ini`
-
-Keyword arguments 
-^^^^^^^^^^^^^^^^^
-Pass the arguments with valid names (listed below). This will override the default values at `configuration.ini`: 
-
-The allowed `kwargs` are::
-
-    kwargs_ = ['ne', 'nu', 'network_type_ee', 'network_type_ei', 'network_type_ie', 
-               'lambda_ee','lambda_ei', 'lambda_ie', 'eta_stdp','eta_inhib', 'eta_ip', 
-               'te_max', 'ti_max', 'ti_min', 'te_min', 'mu_ip','sigma_ip']
-
-Simulation
-----------
-Plasticity Phase
-^^^^^^^^^^^^^^^^
-The default ne, nu values are overriden by passing them as kwargs insidesimulate_sorn method::
-
-    # Import 
     from sorn import Simulator
     import numpy as np
 
-    # Sample input 
+    # Sample input
     num_features = 10
     time_steps = 200
     inputs = np.random.rand(num_features,time_steps)
 
-    # To simulate the network
-    matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=None, noise = True, time_steps=time_steps, ne = 200, nu=num_features)
+    # Simulate the network with default hyperparameters under gaussian white noise
 
-    # To resume the simulation, load the matrices_dict from previous simulation
-    matrices_dict, Exc_activity, Inh_activity, Rec_activity, num_active_connections = Simulator.simulate_sorn(inputs = inputs, phase='plasticity', matrices=matrices_dict, noise= True, time_steps=time_steps,ne = 200, nu=num_features)
+    state_dict, E, I, R, C = Simulator.simulate_sorn(inputs = inputs, phase='plasticity',
+                                                    matrices= None, noise = True,
+                                                    time_steps=time_steps)
 
-Training Phase
-^^^^^^^^^^^^^^
-Train the network using the matrices obtained from plasticity phase::
+The default values of the network hyperparameters are,
 
-    # Import
+.. list-table:: Hyperparameters of the network and default values
+   :widths: 25 25 50
+   :header-rows: 1
+
+   * - Keyword argument
+     - Value
+     - Description
+   * - ne,
+     - 200
+     - Number of Encitatory neurons in the reservoir
+   * - nu
+     - 10
+     - Number of Input neurons in the reservoir
+   * - network_type_ee
+     - "Sparse"
+     - `Sparse` or `Dense` connectivity between Excitatory neurons
+   * - network_type_ie
+     - "Dense"
+     - `Sparse` or `Dense` connectivity from Excitatory to Inhibitory neurons
+   * - _network_type_ei
+     - "Sparse"
+     - `Sparse` or `Dense` connectivity from Inhibitory to Excitatory neurons
+   * - lambda_ee
+     - 20
+     - % of connections between neurons in Excitatory pool
+   * - lambda_ei
+     - 40
+     - % of connections from Inhibitory to Excitatory neurons
+   * - lambda_ie
+     - 100
+     - % of connections from Excitatory to Inhibitory neurons
+   * - eta_stdp
+     - 0.004
+     - Hebbian Learning rate for connections between excitatory neurons
[…]
