+++ docs/build/html/_sources/usage.rst.txt
@@ -2,7 +2,9 @@ Usage
 =====
 
 Plasticity Phase
------------------
+----------------
+
+.. code-block:: python
 
     import sorn
     from sorn import Simulator
@@ -14,12 +16,9 @@ Plasticity Phase
     inputs = np.random.rand(num_features,time_steps)
 
     # Simulate the network with default hyperparameters under gaussian white noise
-
     state_dict, E, I, R, C = Simulator.simulate_sorn(inputs = inputs, phase='plasticity',
                                                     matrices= None, noise = True,
                                                     time_steps=time_steps)
-
-The default values of the network hyperparameters are,
 
 .. list-table:: Hyperparameters of the network and default values
    :widths: 25 25 50
@@ -40,7 +39,7 @@ The default values of the network hyperparameters are,
    * - network_type_ie
      - "Dense"
      - `Sparse` or `Dense` connectivity from Excitatory to Inhibitory neurons
-   * - _network_type_ei
+   * - network_type_ei
      - "Sparse"
      - `Sparse` or `Dense` connectivity from Inhibitory to Excitatory neurons
    * - lambda_ee
@@ -55,7 +54,7 @@ The default values of the network hyperparameters are,
    * - eta_stdp
      - 0.004
      - Hebbian Learning rate for connections between excitatory neurons
-   * - eta_stdp
+   * - eta_inhib
      - 0.001
      - Hebbian Learning rate for connections from Inhibitory to Excitatory neurons
    * - eta_ip
@@ -80,7 +79,9 @@ The default values of the network hyperparameters are,
      - 0.0
      - Standard deviation of firing rate of excitatory neuron
 
-To override the default hyperparameters, use the `kwargs` as shown below,
+To override the default hyperparameters, use the `kwargs` as shown below:
+
+.. code-block:: python
 
     state_dict, E, I, R, C = Simulator.simulate_sorn(inputs = inputs, phase='plasticity',
                                                     matrices=None, noise= True,
@@ -88,15 +89,16 @@ To override the default hyperparameters, use the `kwargs` as shown below,
                                                     ne = 200, nu=num_features)
 
 Training phase
----------------
+--------------
+
+.. code-block:: python
 
     from sorn import Trainer
-    # NOTE: During training phase, input to `sorn` should have second (time) dimension set to 1. ie., input shape should be (input_features,1).
-
+    # NOTE: During training phase, input to `sorn` should have second (time) dimension set to 1. ie., input shape
+    # should be (input_features,1).
     inputs = np.random.rand(num_features,1)
 
     # SORN network is frozen during training phase
-
     state_dict, E, I, R, C = Trainer.train_sorn(inputs = inputs, phase='training',
                                                 matrices=state_dict,
                                                 nu=num_features, time_steps=1)
@@ -105,7 +107,9 @@ Freeze plasticity
 -----------------
 
 To turn off any plasticity mechanisms during `simulation` or `training` phase, use `freeze` argument.
-For example to stop intrinsic plasticity during simulation phase,
+For example to stop intrinsic plasticity during simulation phase
+
+.. code-block:: python
 
     # Sample input
     num_features = 10
@@ -119,45 +123,46 @@ For example to stop intrinsic plasticity during simulation phase,
 
 The other options are,
 
[…]
