### 2021-08-15 265bffb
#31 v0.5.6 config removed, docs, readme, openai example updates
