### 2021-08-15 6d9231b
#31 v0.5.6 config removed, docs, readme, openai example updates
