+++ sorn/sorn.py
@@ -17,7 +17,7 @@ class Sorn(object):
     Args:
         nu (int): Number of input units. Defaults to 10
 
-        ne (int): Sorn._ne  # Number of excitatory units. Defaults to 200
+        ne (int): Number of excitatory units. Defaults to 200
 
         eta_stdp (float): STDP plasticity Learning rate constant; SORN1 and SORN2. Defaults to 0.004
 
@@ -25,13 +25,13 @@ class Sorn(object):
 
         eta_inhib (float): Intrinsic plasticity learning rate constant; SORN2 only. Defaults to 0.01
 
-        h_ip (float): Target firing rate. Defaults to 2 * Sorn._nu / Sorn._ne
+        h_ip (float): Target firing rate. Defaults to 2 * nu / ne
 
         mu_ip (float): Mean target firing rate. Defaults to 0.1
 
         sigma_ip (float): Variance of target firing rate. Defaults to 0.0
 
-        ni (int): Number of inhibitory units in the network. Defaults to int(0.2 * Sorn._ne)
+        ni (int): Number of inhibitory units in the network. Defaults to int(0.2*ne)
 
         time_steps (float): Total time steps of simulation
 
@@ -56,37 +56,36 @@ class Sorn(object):
         lambda_ie (int):  Number of connections to and from a single excitatory unit to inhibitory unit at initialization. Defaults to 100"""
 
     def __init__(self):
+        pass
 
-        self.nu = 10
-        self.ne = 200
-        self.ni = int(0.2 * self.ne)
-        self.eta_stdp = 0.004
-        self.eta_inhib = 0.001
-        self.eta_ip = 0.01
-        self.te_max = 1.0
-        self.ti_max = 0.5
-        self.ti_min = 0.0
-        self.te_min = 0.0
-        self.mu_ip = 0.1
-        self.sigma_ip = 0.0
-        self.network_type_ee = "Sparse"
-        self.network_type_ei = "Sparse"
-        self.network_type_ie = "Dense"
-        self.lambda_ee = 20
-        self.lambda_ei = 40
-        self.lambda_ie = 100
-
-    def initialize_weight_matrix(self,
+    nu = 10
+    ne = 200
+    ni = int(0.2 * ne)
+    eta_stdp = 0.004
+    eta_inhib = 0.001
+    eta_ip = 0.01
+    te_max = 1.0
+    ti_max = 0.5
+    ti_min = 0.0
+    te_min = 0.0
+    mu_ip = 0.1
+    sigma_ip = 0.0
+    network_type_ee = "Sparse"
+    network_type_ei = "Sparse"
+    network_type_ie = "Dense"
+    lambda_ee = 20
+    lambda_ei = 40
+    lambda_ie = 100
+    @staticmethod
+    def initialize_weight_matrix(
         network_type: str, synaptic_connection: str, self_connection: str, lambd_w: int
     ):
         """Wrapper for initializing the weight matrices for SORN
-
         Args:
             network_type (str): Spare or Dense
             synaptic_connection (str): EE,EI,IE. Note that Spare connection is defined only for EE connections
             self_connection (str): True or False: Synaptic delay or time delay
             lambd_w (int): Average number of incoming and outgoing connections per neuron
-
         Returns:
             weight_matrix (array): Array of connection strengths
         """
@@ -94,109 +93,103 @@ class Sorn(object):
         if (network_type == "Sparse") and (self_connection == "False"):
 
             # Generate weight matrix for E-E/ E-I connections with mean lamda incoming and out-going connections per neuron
-            assert (lambd_w <= self.ne), "Number of connections per unit (lambda) should be less than number of units(Ne) in the pool and also Ne should be greater than 25"
[…]
