+++ docs/source/installation.rst
@@ -1,19 +1,17 @@
 Installation
 ============
 
-Install using pip:
+Install using `pip`
+
+.. code-block:: python
 
    pip install sorn
 
 or
 
-using conda:
+To install the latest version from the development branch
 
-    conda install -c conda-forge sorn
-
-or
-
-The library is still in alpha stage, so you may want to install the latest version from the development branch:
+.. code-block:: python
 
     pip install git+https://github.com/Saran-nns/sorn
 
@@ -23,7 +21,9 @@ Dependencies
 
 SORN supports Python 3.5+ ONLY. For older Python versions please use the official Python client
 
-To install all optional dependencies run:
+To install all optional dependencies run
 
-  pip install 'sorn[all]'
+.. code-block:: python
+
+    pip install 'sorn[all]'
 
