+++ build/lib/sorn/__init__.py
@@ -3,6 +3,6 @@ import logging
 from .utils import *
 
 __author__ = "Saranraj Nambusubramaniyan"
-__version__ = "0.5.6"
+__version__ = "0.5.7"
 
 logging.basicConfig(level=logging.INFO)
