+++ .readthedocs.yml
@@ -13,7 +13,7 @@ sphinx:
 formats: all
 
 python:
-  version: 3.8
+  version: 3.6
   install:
     - requirements: docs/requirements.txt
     - method: pip
