+++ sorn.egg-info/requires.txt
@@ -2,4 +2,5 @@ numpy
 configparser
 scipy
 seaborn
+pandas
 networkx
