+++ .github/workflows/main.yml
@@ -1,0 +1,29 @@
+name: AutomaticHost
+on:
+  push:
+    branches:
+      - { master }
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-node@v1
+        with:
+          node-version: "10.x"
+      - run: |
+          npm ci
+          {BUILD_COMMAND}
+      - run: |
+          git config --local user.email "action@github.com"
+          git config --local user.name "GitHub Action"
+          git checkout --orphan gh-pages
+          git reset
+          git add {OUT_DIR}/* -f
+          git mv {OUT_DIR}/* ./ -k
+          git commit -m "Update hosted version"
+      - uses: ad-m/github-push-action@master
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          force: true
+          branch: gh-pages
