+++ .github/workflows/build.yml
@@ -1,0 +1,28 @@
+name: Build and Test [Python 3.6, 3,7, 3,8]
+
+on: [push, pull_request]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.5, 3.6, 3.7, 3.8]
+    steps:
+      - uses: actions/checkout@v2
+      - name: Build using Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{matrix.python-version }}
+      - name: install dependencies [pip]
+        run: |
+          python -m pip install pip --upgrade pip
+          pip install flake8
+          if [ -f requirements.txt ]; then pip install requirements.txt; fi
+
+          pip install -e .
+      - name: run linter [flake8]
+        run: flake8
+
+      - name: run unit tests [pytest]
+        run: unittest
