+++ .github/workflows/build.yml
@@ -18,11 +18,10 @@ jobs:
         run: |
           python -m pip install pip --upgrade pip
           pip install flake8
-          if [ -f requirements.txt ]; then pip install requirements.txt; fi
-
+          pip install -r requirements.txt
           pip install -e .
       - name: run linter [flake8]
         run: flake8
 
-      - name: run unit tests [pytest]
-        run: unittest
+      - name: Unit tests
+        run: python -m unittest test_sorn.py
