+++ requirements.txt
@@ -1,7 +1,7 @@
 certifi==2021.5.30
 configparser==5.0.2
 cycler==0.10.0
-networkx==2.6.2
+networkx
 kiwisolver==1.3.1
 matplotlib==3.4.2
 numpy==1.21.1
