+++ requirements.txt
@@ -3,14 +3,14 @@ configparser==5.0.2
 cycler==0.10.0
 networkx
 kiwisolver==1.3.1
-matplotlib==3.4.2
-numpy==1.21.1
-pandas==1.3.1
+matplotlib
+numpy
+pandas
 pyparsing==2.4.7
 python-dateutil==2.8.2
 pytz==2021.1
-scipy==1.7.1
-seaborn==0.11.1
+scipy
+seaborn
 six==1.16.0
 wincertstore==0.2
 
