+++ requirements.txt
@@ -1,16 +1,16 @@
-certifi==2021.5.30
-configparser==5.0.2
-cycler==0.10.0
+certifi
+configparser
+cycler
 networkx
-kiwisolver==1.3.1
+kiwisolver
 matplotlib
 numpy
 pandas
-pyparsing==2.4.7
-python-dateutil==2.8.2
-pytz==2021.1
+pyparsing
+python-dateutil
+pytz
 scipy
 seaborn
-six==1.16.0
-wincertstore==0.2
+six
+wincertstore
 
