+++ .github/workflows/build.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.5, 3.6, 3.7, 3.8]
+        python-version: [3.5, 3.7, 3.8]
     steps:
       - uses: actions/checkout@v2
       - name: Build using Python ${{ matrix.python-version }}
