+++ .github/workflows/build.yml
@@ -20,8 +20,8 @@ jobs:
           pip install flake8
           pip install -r requirements.txt
           pip install -e .
-      - name: run linter [flake8]
-        run: flake8
+      # - name: run linter [flake8]
+      #   run: flake8
 
       - name: Unit tests
         run: python -m unittest test_sorn.py
