+++ .github/workflows/build.yml
@@ -1,4 +1,4 @@
-name: Build and Test [Python 3.6, 3,7, 3,8]
+name: Build and Test [Python 3.5, 3.6, 3,7, 3,8, 3.9]
 
 on: [push, pull_request]
 
