+++ .github/workflows/build.yml
@@ -1,4 +1,4 @@
-name: TestBuild
+name: Build
 
 on: [push, pull_request]
 
