### 2021-08-22 a67fbf3
Delete main.yml
