+++ .github/workflows/build.yml
@@ -20,8 +20,25 @@ jobs:
           pip install flake8
           pip install -r requirements.txt
           pip install -e .
-      # - name: run linter [flake8]
-      #   run: flake8
 
       - name: Unit tests
         run: python -m unittest test_sorn.py
+
+      - name: Generate coverage report
+        run: |
+          pip install pytest
+          pip install pytest-cov
+          pytest --cov=./ --cov-report=xml
+
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v2
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          directory: ./coverage/reports/
+          env_vars: OS,PYTHON
+          fail_ci_if_error: true
+          files: ./coverage1.xml,./coverage2.xml
+          flags: unittests
+          name: codecov-umbrella
+          path_to_write_report: ./coverage/codecov_report.txt
+          verbose: true
