+++ .github/workflows/build.yml
@@ -18,6 +18,7 @@ jobs:
         run: |
           python -m pip install pip --upgrade pip
           pip install flake8
+          pip install coverage
           pip install -r requirements.txt
           pip install -e .
 
