+++ .github/workflows/build.yml
@@ -18,24 +18,22 @@ jobs:
         run: |
           python -m pip install pip --upgrade pip
           pip install flake8
-          pip install coverage
+          pip install pytest
           pip install -r requirements.txt
           pip install -e .
 
       - name: Unit tests
         run: |
           python -m unittest test_sorn.py
-          coverage run test_sorn.py
+
+          py.test . --cov-report=xml --cov=sorn -vvv
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v2
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
-          directory: ./coverage/reports/
           env_vars: OS,PYTHON
-          fail_ci_if_error: true
-          files: ./coverage1.xml,./coverage2.xml
           flags: unittests
           name: codecov-umbrella
-          path_to_write_report: ./coverage/codecov_report.txt
-          verbose: true
+          fail_ci_if_error: false
+          verbose: false
