+++ .github/workflows/build.yml
@@ -26,8 +26,6 @@ jobs:
         run: |
           python -m unittest test_sorn.py
 
-          py.test . --cov-report=xml --cov=sorn -vvv
-
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v2
         with:
