+++ .github/workflows/build.yml
@@ -28,4 +28,4 @@ jobs:
           python -m unittest test_sorn.py
           pytest --cov=./ --cov-report=xml
 
-        uses: codecov/codecov-action@v2
+      - uses: codecov/codecov-action@v2
