+++ .github/workflows/build.yml
@@ -19,7 +19,7 @@ jobs:
           python -m pip install pip --upgrade pip
           pip install flake8
           pip install pytest
-          pip install coverage
+          pip install pytest-cov
           pip install -r requirements.txt
           pip install -e .
 
