+++ .github/workflows/build.yml
@@ -26,6 +26,6 @@ jobs:
       - name: Unit tests
         run: |
           python -m unittest test_sorn.py
-          pytest --cov=./ --cov-report=xml
+          py.test --cov=./ --cov-report=xml
 
       - uses: codecov/codecov-action@v2
