+++ .github/workflows/build.yml
@@ -26,6 +26,9 @@ jobs:
       - name: Unit tests
         run: |
           python -m unittest test_sorn.py
+
+      - name: Coverage
+        run: |
           py.test --cov=./ --cov-report=xml
 
       - uses: codecov/codecov-action@v2
