+++ .github/workflows/build.yml
@@ -2,9 +2,6 @@ name: Build
 
 on: [push, pull_request]
 
-codecov:
-  disable_default_path_fixes: true
-
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -20,18 +17,15 @@ jobs:
       - name: install dependencies [pip]
         run: |
           python -m pip install pip --upgrade pip
-          pip install flake8
-          pip install pytest
-          pip install pytest-cov
+          pip install coverage
           pip install -r requirements.txt
           pip install -e .
 
       - name: Unit tests
-        run: |
-          python -m unittest test_sorn.py
+        run: python -m unittest test_sorn.py
 
-      - name: Coverage
-        run: |
-          py.test --cov=./ --cov-report=xml
-
-      - uses: codecov/codecov-action@v2
+      - name: Report coverage to codecov
+          run:
+            coverage run test_sorn.py.py test && coverage report --fail-under=75 && coverage xml
+          after_success:
+            - codecov
