+++ paper/paper.bib
@@ -20,6 +20,22 @@
   publisher={Public Library of Science San Francisco, USA}
 }
 
+@article{jaeger2002adaptive,
+  title={Adaptive nonlinear system identification with echo state networks},
+  author={Jaeger, Herbert},
+  journal={Advances in neural information processing systems},
+  volume={15},
+  pages={609--616},
+  year={2002}
+}
+
+@article{jaeger2007special,
+  title={Special issue on echo state networks and liquid state machines.},
+  author={Jaeger, Herbert and Maass, Wolfgang and Principe, Jose},
+  year={2007},
+  publisher={Elsevier Science}
+}
+
 @article{papa2017criticality,
   title={Criticality meets learning: Criticality signatures in a self-organizing recurrent neural network},
   author={Papa, Bruno Del and Priesemann, Viola and Triesch, Jochen},
