+++ test_sorn.py
@@ -7,7 +7,7 @@ from sorn.utils import Plotter, Statistics, Initializer
 # Getting back the pickled matrices:
 with open("sample_matrices.pkl", "rb") as f:
     (
-        matrices_dict,
+        state_dict,
         Exc_activity,
         Inh_activity,
         Rec_activity,
@@ -27,6 +27,11 @@ inputs = np.random.rand(num_features, time_steps)
 
 class TestSorn(unittest.TestCase):
     def test_runsorn(self):
+        """Test network initialization, simulation and training with different
+        initialization methods and plasticity rules
+        """
+
+        # Initialize and simulate SORN with the default hyperparameters
         self.assertRaises(
             Exception,
             Simulator.simulate_sorn(
@@ -38,66 +43,65 @@ class TestSorn(unittest.TestCase):
                 nu=num_features
             ),
         )
+        # Initilize and resume the simulation of SORN using the state dictionary, state_dict
         self.assertRaises(
             Exception,
             Simulator.simulate_sorn(
                 inputs=simulation_inputs,
                 phase="plasticity",
-                matrices=matrices_dict,
+                matrices=state_dict,
                 time_steps=2,
                 noise=False,
             ),
         )
+        # Freeze a particular plasticity during simulation
         self.assertRaises(
             Exception,
             Simulator.simulate_sorn(
                 inputs=simulation_inputs,
                 phase="plasticity",
-                matrices=matrices_dict,
+                matrices=state_dict,
                 time_steps=2,
                 noise=False, freeze=['ip']
             ),
         )
+
+        # Freeze multiple plasticity mechanisms during simulation
         self.assertRaises(
             Exception,
             Simulator.simulate_sorn(
                 inputs=simulation_inputs,
                 phase="plasticity",
-                matrices=matrices_dict,
+                matrices=state_dict,
                 time_steps=2,
-                noise=False, freeze=['stdp,istdp','ss','sp']
+                noise=False, freeze=['stdp','istdp','ss','sp']
             ),
         )
+
+        # Train SORN with all plasticity mechanisms active
         self.assertRaises(
             Exception,
             Trainer.train_sorn(
                 inputs=gym_input,
                 phase="plasticity",
-                matrices=matrices_dict,
+                matrices=state_dict,
                 time_steps=1,
                 noise=True,
             ),
         )
-        self.assertRaises(
-            Exception,
-            Trainer.train_sorn(
-                inputs=sequence_input,
-                phase="plasticity",
-                matrices=matrices_dict,
-                time_steps=1,
-                noise=True,
-            ),
-        )
+        # Freeze multiple plasticity mechanisms during training
         self.assertRaises(
             Exception,
             Trainer.train_sorn(
                 inputs=sequence_input,
                 phase="training",
-                matrices=matrices_dict,
+                matrices=state_dict,
                 time_steps=1,
-                noise=True, freeze=['stdp,istdp','ss','sp']
+                noise=True, freeze=['stdp','istdp','ss','sp']
             ),
         )
+
+        # Override the default hyperparameters, initialize SORN and simulate under all plasticity mechanisms
         self.assertRaises(
             Exception,
             Simulator.simulate_sorn(
@@ -113,6 +117,7 @@ class TestSorn(unittest.TestCase):
             ),
         )
 
+        # Override the default hyperparameters, initialize SORN and train under all plasticity mechanisms
         self.assertRaises(
             Exception,
             Trainer.train_sorn(
@@ -129,18 +134,24 @@ class TestSorn(unittest.TestCase):
         )
 
     def test_plotter(self):
+        """Test the Plotter class methods in utils module
+        """
+
[…]
