+++ docs/source/usage.rst
@@ -276,7 +276,7 @@ Plot weight distribution in the network
     Wee=Wee/Wee.max() # state_dict['Wee'] returned by the SORN is already normalized
     Plotter.weight_distribution(weights= Wee, bin_size = 5, savefig = True)
 
-.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/revision/imgs/weight_distribution.png
+.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/master/imgs/weight_distribution.png
 
 
 Plot Spike train
@@ -287,7 +287,7 @@ Plot Spike train
     E = np.random.randint(2, size=(200,1000)) # For example, activity of 200 excitatory neurons in 1000 time steps
     Plotter.scatter_plot(spike_train = E, savefig=True)
 
-.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/revision/imgs/ScatterSpikeTrain.png
+.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/master/imgs/ScatterSpikeTrain.png
 
 
 Raster plot of Spike train
@@ -297,7 +297,7 @@ Raster plot of Spike train
     # Raster plot of activity of only first 10 neurons in the excitatory pool
     Plotter.raster_plot(spike_train = E[:,0:10], savefig=True)
 
-.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/revision/imgs/RasterSpikeTrain.png
+.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/master/imgs/RasterSpikeTrain.png
 
 Distribution of presynaptic connections
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -306,7 +306,7 @@ Distribution of presynaptic connections
     # Histogram of number of presynaptic connections per neuron in the excitatory pool
     Plotter.hist_incoming_conn(weights=Wee, bin_size=10, histtype='bar', savefig=True)
 
-.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/revision/imgs/hist_incoming_conn.png
+.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/master/imgs/hist_incoming_conn.png
 
 
 Distribution of firing rate of the network
@@ -315,7 +315,7 @@ Distribution of firing rate of the network
 
     Plotter.hist_firing_rate_network(E, bin_size=10, savefig=True)
 
-.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/revision/imgs/hist_firing_rate_network.png
+.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/master/imgs/hist_firing_rate_network.png
 
 Plot pearson correlation between neurons
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -325,7 +325,7 @@ Plot pearson correlation between neurons
     avg_corr_coeff,_ = Statistics.avg_corr_coeff(E)
     Plotter.correlation(avg_corr_coeff,savefig=True)
 
-.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/revision/imgs/correlation_between_neurons.png
+.. image:: https://raw.githubusercontent.com/Saran-nns/sorn/master/imgs/correlation_between_neurons.png
 
 Inter spike intervals
 ^^^^^^^^^^^^^^^^^^^^^^
@@ -334,7 +334,7 @@ Inter spike intervals
     # Inter spike intervals with exponential curve fit for neuron 1 in the Excitatory pool
     Plotter.isi_exponential_fit(E,neuron=1,bin_size=10, savefig=True)
 
[…]
