+++ docs/source/usage.rst
@@ -17,8 +17,16 @@ Plasticity Phase
 
     # Simulate the network with default hyperparameters under gaussian white noise
     state_dict, E, I, R, C = Simulator.simulate_sorn(inputs = inputs, phase='plasticity',
-                                                    matrices= None, noise = True,
+                                                    matrices=None, noise = True,
                                                     time_steps=time_steps)
+
+.. code-block:: python
+
+    Network Initialized
+    Number of connections in Wee 3909 , Wei 1574, Wie 8000
+    Shapes Wee (200, 200) Wei (40, 200) Wie (200, 40)
+
+The default values of the network hyperparameters are,
 
 .. list-table:: Hyperparameters of the network and default values
    :widths: 25 25 50
@@ -79,14 +87,27 @@ Plasticity Phase
      - 0.0
      - Standard deviation of firing rate of excitatory neuron
 
-To override the default hyperparameters, use the `kwargs` as shown below:
+Override the default hyperparameters and simulate new SORN model
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. code-block:: python
 
+    # Sample input
+    num_features = 5
+    time_steps = 1000
+    inputs = np.random.rand(num_features,time_steps)
+
     state_dict, E, I, R, C = Simulator.simulate_sorn(inputs = inputs, phase='plasticity',
-                                                    matrices=None, noise= True,
+                                                    matrices=None, noise = True,
                                                     time_steps=time_steps,
-                                                    ne = 200, nu=num_features)
+                                                    ne = 100, nu=num_features,
+                                                    lambda_ee = 10, eta_stdp=0.001)
+
+.. code-block:: python
+
+    Network Initialized
+    Number of connections in Wee 959 , Wei 797, Wie 2000
+    Shapes Wee (100, 100) Wei (20, 100) Wie (100, 20)
 
 Training phase
 --------------
@@ -94,32 +115,53 @@ Training phase
 .. code-block:: python
 
     from sorn import Trainer
-    # NOTE: During training phase, input to `sorn` should have second (time) dimension set to 1. ie., input shape
-    # should be (input_features,1).
+    # NOTE: During training phase, input to `sorn` should have second (time) dimension set to 1. ie., input shape should be (input_features,1).
+
     inputs = np.random.rand(num_features,1)
 
     # SORN network is frozen during training phase
     state_dict, E, I, R, C = Trainer.train_sorn(inputs = inputs, phase='training',
-                                                matrices=state_dict,
-                                                nu=num_features, time_steps=1)
+                                                matrices=state_dict, noise= False,
+                                                time_steps=1,
+                                                ne = 100, nu=num_features,
+                                                lambda_ee = 10, eta_stdp=0.001 )
 
 Freeze plasticity
 -----------------
 
-To turn off any plasticity mechanisms during `simulation` or `training` phase, use `freeze` argument.
-For example to stop intrinsic plasticity during simulation phase
+To turn off any plasticity mechanisms during simulation or training phase, use freeze argument.
+For example to stop intrinsic plasticity during simulation phase,
 
 .. code-block:: python
 
     # Sample input
     num_features = 10
-    time_steps = 200
+    time_steps = 20
     inputs = np.random.rand(num_features,time_steps)
 
     state_dict, E, I, R, C = Simulator.simulate_sorn(inputs = inputs, phase='plasticity',
                                                     matrices=None, noise = True,
-                                                    time_steps=time_steps, ne = 200,
+                                                    time_steps=time_steps, ne = 50,
[…]
