+++ README.md
@@ -4,6 +4,7 @@
 SORN is a class of neuro-inspired artificial network build based on plasticity mechanisms in biological brain and mimic neocortical circuits ability of learning and adaptation. SORN consists of pool of excitatory neurons and small population of inhibitory neurons which are controlled by 5 plasticity mechanisms found in neocortex, namely Spike Ti ming Dependent Plasticity (STDP), Intrinsic Plasticity (IP), Synaptic Scaling (SS),Synaptic Normalization(SN) and inhibitory Spike Timing Dependent Plasticity (iSTDP).
 
 'sorn' is a Python package designed for Self Organizing Recurrent Neural Networks. It encapsulates all 5 plastcity mechansims and allows researchers to develop the network of their interest, provided they have the freedom to choose the combination of plasticity rules of their choice While it was originally developed for SORN networks, it can also serve as an ideal research package for Liquid State Machines.
+
 [![Build Status](https://github.com/saran-nns/sorn/workflows/Build/badge.svg)](https://github.com/saran-nns/sorn/actions)
 [![codecov](https://codecov.io/gh/Saran-nns/sorn/branch/master/graph/badge.svg)](https://codecov.io/gh/Saran-nns/sorn)
 [![Documentation Status](https://readthedocs.org/projects/self-organizing-recurrent-neural-networks/badge/?version=latest)](https://self-organizing-recurrent-neural-networks.readthedocs.io/en/latest/?badge=latest)
@@ -56,8 +57,6 @@ To install all optional dependencies,
 ```python
   pip install 'sorn[all]'
 ```
-For detailed documentation about usage and development, please visit [SORN-Documentation](https://self-organizing-recurrent-neural-networks.readthedocs.io/)
-
 ## Usage
 ### Plasticity Phase
 
@@ -82,116 +81,20 @@ Network Initialized
 Number of connections in Wee 3909 , Wei 1574, Wie 8000
 Shapes Wee (200, 200) Wei (40, 200) Wie (200, 40)
 ```
-
-The default values of the network hyperparameters are,
-
-Keyword argument | Value | Description |
---- | --- | --- |
-ne              | 200      | Number of Encitatory neurons in the reservoir                              |
-nu              | 10       | Number of Input neurons in the reservoir                                   |
-network_type_ee | "Sparse" | `Sparse` or `Dense` connectivity between Excitatory neurons                |
-network_type_ie | "Dense"  | `Sparse` or `Dense` connectivity from Excitatory to Inhibitory neurons     |
-network_type_ei | "Sparse" | `Sparse` or `Dense` connectivity from Inhibitory to Excitatory neurons     |
-lambda_ee       | 20       | % of connections between neurons in Excitatory pool                        |
-lambda_ei       | 40       | % of connections from Inhibitory to Excitatory neurons                     |
-lambda_ie       | 100      | % of connections from Excitatory to Inhibitory neurons                     |
-eta_stdp        | 0.004    | Hebbian Learning rate for connections between excitatory neurons           |
-eta_inhib       | 0.001    | Hebbian Learning rate for connections from Inhibitory to Excitatory neurons|
-eta_ip          | 0.01     | Intrinsic plasticity learning rate                                         |
-te_max          | 1.0      | Maximum excitatory neuron threshold value                                  |
-ti_max          | 0.5      | Maximum inhibitory neuron threshold value                                  |
-ti_min          | 0.0      | Minimum inhibitory neuron threshold value                                  |
-te_min          | 0.0      | Minimum excitatory neuron threshold value                                  |
[…]
