### 2021-09-07 92a2c7a
refactor sorn.py
