+++ sorn/sorn.py
@@ -12,51 +12,7 @@ except:
 
 class Sorn(object):
 
-    """ This class wraps initialization of the network and its parameters
-
-    Args:
-        nu (int): Number of input units. Defaults to 10
-
-        ne (int): Number of excitatory units. Defaults to 200
-
-        eta_stdp (float): STDP plasticity Learning rate constant; SORN1 and SORN2. Defaults to 0.004
-
-        eta_ip (float): Intrinsic plasticity learning rate constant; SORN1 and SORN2. Defaults to 0.001
-
-        eta_inhib (float): Intrinsic plasticity learning rate constant; SORN2 only. Defaults to 0.01
-
-        h_ip (float): Target firing rate. Defaults to 2 * nu / ne
-
-        mu_ip (float): Mean target firing rate. Defaults to 0.1
-
-        sigma_ip (float): Variance of target firing rate. Defaults to 0.0
-
-        ni (int): Number of inhibitory units in the network. Defaults to int(0.2*ne)
-
-        time_steps (float): Total time steps of simulation
-
-        te_min (float): Excitatory minimum Threshold. Defaults to 0.0
-
-        te_max (float): Excitatory maximum Threshold. Defaults to 1.0
-
-        ti_min (float): Inhibitory minimum Threshold. Defaults to 0.0
-
-        ti_max (float): Inhibitory maximum Threshold. Defaults to 0.5
-
-        network_type_ee (str): Dense or Sparse. Defaults to Sparse
-
-        network_type_ei (str): Dense or Sparse. Defaults to Sparse
-
-        network_type_ie (str): Dense or Sparse. Defaults to Dense
-
-        lambda_ee (int):  Number of connections to and from a single excitatory unit to another at initialization. Defaults to 20
-
-        lambda_ei (int): Number of connections to and from a single inhibitory unit to exitatory unit at initialization. Defaults to 40
-
-        lambda_ie (int):  Number of connections to and from a single excitatory unit to inhibitory unit at initialization. Defaults to 100"""
-
-    def __init__(self):
-        pass
+    """ This class wraps initialization of the network and its parameters"""
 
     nu = 10
 
@@ -95,15 +51,19 @@ class Sorn(object):
     lambda_ie = 100
 
     @staticmethod
-    def initialize_weight_matrix(
-        network_type: str, synaptic_connection: str, self_connection: str, lambd_w: int
+    def initialize_weight_matrix(network_type: str, synaptic_connection: str, self_connection: str, lambd_w: int
     ):
         """Wrapper for initializing the weight matrices for SORN
+
         Args:
             network_type (str): Spare or Dense
+
             synaptic_connection (str): EE,EI,IE. Note that Spare connection is defined only for EE connections
+
             self_connection (str): True or False: Synaptic delay or time delay
+
             lambd_w (int): Average number of incoming and outgoing connections per neuron
+
         Returns:
             weight_matrix (array): Array of connection strengths
         """
@@ -126,15 +86,15 @@ class Sorn(object):
         return weight_matrix
 
     @staticmethod
-    def initialize_threshold_matrix(
-        te_min: float, te_max: float, ti_min: float, ti_max: float
-    ):
+    def initialize_threshold_matrix(te_min: float, te_max: float, ti_min: float, ti_max: float):
         """Initialize the threshold for excitatory and inhibitory neurons
+
         Args:
             te_min (float): Min threshold value for excitatory units
             te_max (float): Min threshold value for inhibitory units
             ti_min (float): Max threshold value for excitatory units
             ti_max (float): Max threshold value for inhibitory units
+
         Returns:
             te (array): Threshold values for excitatory units
[…]
