### 2021-09-07 #3545 @rkurchin
:tada: Congrats @Saran-nns! Now I get to do the "editing" part of my editor job. Here are a few small editorial/stylistic tweaks for the manuscript:
* line 7: need space between "neural" and "(SORN)"
* 11: need another comma after "namely"
* 14: insert "a" between "tools," and "SORN"
* 17: instead of 'sorn', I suggest you put `sorn` with backticks (`) to get monospace typesetting (change in every instance)
* 21: feels like a word is missing, perhaps "...applications of this network architecture, ..." would be better?
* 25: change beginning to "Reservoir computing (RC) models..." to define acronym
* 38: should read "...for SORN networks"
* 39: remove "considered," consider rephrasing to "application-specific" (and insert hyphen), insert hyphen in "general-purpose"
* 41: rephrase, "providing them the freedom to choose..."
* 45: I suggest inserting hyphen in "reverse-engineering," though the internet seems split on this usage
* "Library Overview" paragraph: again, would suggest backticks for all package names and code quotations
* 49: remove the word "roughly," it's either five or it's not five (also spell out the word "five" rather than using the numeral)
* 56: insert colon after "two classes" and comma after `Trainer`
* equation typesetting in general: when there are superscripts/subscripts that are not variables/indices (i.e. all the EE, EI, IE, IP, etc.), they should be in `\text{}` rather than the mathematical italics
* 71: unclear what "It" is referring to
* equation 3: missing a right parenthesis
* 86: again, I assume "It" refers to structural plasticity, but would be good to spell it out
* 89: define acronym SS if you're going to use it
* 92: define acronym
* equation 6: crossed parentheses pairs
* 95: typo: "strenght"
* 98: comma after "that" is unnecessary
* 101: suggest replacing "last" with "previous"
* table after line 126: I suggest widening left column so that each kwarg fits on one line
* 128: Insert "The" before `sorn`
* 128-129: rephrase sentence to "A few of the methods in the `Statistics` module are:" (and again, widen left column in ensuing table)
* 130: replace "is" with "are" or "can be"
