+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 # Summary
 
 
-The self-organizing recurrent neural (SORN) network is a class of neuro-inspired artificial networks. This class of networks has been shown to mimic the ability of neocortical circuits to learn and adapt through neuroplasticity mechanisms. Structurally, the SORN network consists of a pool of excitatory neurons and a small population of inhibitory neurons. The network uses five basic plasticity mechanisms found in the neocortex of the brain, namely, spike-timing-dependent plasticity, intrinsic plasticity, synaptic scaling, inhibitory spike-timing-dependent plasticity, and structural plasticity [@zheng2013network; @lazar2009sorn; @papa2017criticality] to optimize its parameters. Using a mathematical tools, SORN network simplifies the underlying structural and functional connectivity mechanisms responsible for learning and memory in the brain.
+The self-organizing recurrent neural (SORN) network is a class of neuro-inspired artificial networks. This class of networks has been shown to mimic the ability of neocortical circuits to learn and adapt through neuroplasticity mechanisms. Structurally, the SORN network consists of a pool of excitatory neurons and a small population of inhibitory neurons. The network uses five basic plasticity mechanisms found in the neocortex of the brain, namely, spike-timing-dependent plasticity, intrinsic plasticity, synaptic scaling, inhibitory spike-timing-dependent plasticity, and structural plasticity [@zheng2013network; @lazar2009sorn; @papa2017criticality] to optimize its parameters. Using mathematical tools, a SORN network simplifies the underlying structural and functional connectivity mechanisms responsible for learning and memory in the brain.
 
 ![SORN network](../imgs/SORN1.png){width=80%}
 
@@ -42,7 +42,7 @@ There are few other open source codes [sorn v1](https://github.com/delpapa/SORN)
 
 ## Library Overview
 
-The package `sorn` is heavily dependent on numpy [@harris2020array] for numerical computation and analysis methods, seaborn and matplotlib [@barrett2005matplotlib] for visualization. The network is constructed in five classes; the object `SORN` encapsulates all the required functions that instantiate network variables such as connection weights and thresholds. `Plasticity` inherits objects from `SORN` and implements plasticity rules with methods `stdp()`, `ip()`, `ss()`, `sp()` and `istdp()` . `NetworkState` has methods that evaluate excitatory and inhibitory network states at each time step and finally `MatrixCollection` objects behave like a memory cache. It collects the network states and keeps track of variables such as weights and thresholds as the network evolves during simulation and training.
[…]
