+++ paper/paper.md
@@ -35,9 +35,9 @@ The self-organizing recurrent neural (SORN) network is a class of neuro-inspired
 
 ## Statement of need
 
-Reservoir computing (RC) models are neuroinspired artificial neural networks. RC networks have either sparsely or densely connected units with fixed connection weights. Unlike other RC models, SORN has synaptic weights controlled by neuroinspired plasticity mechanisms. The network has two distinct pools of excitatory and inhibitory reservoirs that compete to remain in a subcritical state suitable for learning. The subcritical state is a state between chaos and order, also called the "edge of chaos". In this state, the network has momentum with a strong affinity for order, but is sensitive to external perturbations. Through plasticity mechanisms, the network has the ability to overcome the perturbations and return to its subcritical dynamics. This self-adaptive behavior is also referred to as self-organization. To build such a network with a synergistic combination of plasticity mechanisms from scratch requires a deeper understanding of neurophysiology and soft computing. sorn' reduces the cognitive load of theorists, experimenters or researchers by encapsulating all plasticity mechanisms with a high degree of reliability and flexibility.
+Reservoir computing (RC) models are neuroinspired artificial neural networks. RC networks have either sparsely or densely connected units with fixed connection weights. Unlike other RC models, SORN has synaptic weights controlled by neuroinspired plasticity mechanisms. The network has two distinct pools of excitatory and inhibitory reservoirs that compete to remain in a subcritical state suitable for learning. The subcritical state is a state between chaos and order, also called the "edge of chaos". In this state, the network has momentum with a strong affinity for order, but is sensitive to external perturbations. Through plasticity mechanisms, the network has the ability to overcome the perturbations and return to its subcritical dynamics. This self-adaptive behavior is also referred to as self-organization. To build such a network with a synergistic combination of plasticity mechanisms from scratch requires a deeper understanding of neurophysiology and soft computing. `sorn` reduces the cognitive load of theorists, experimenters or researchers by encapsulating all plasticity mechanisms with a high degree of reliability and flexibility.
 
-There are few other open source codes [sorn v1](https://github.com/delpapa/SORN), [sorn v2](https://github.com/delpapa/SORN_V2), for SORN networks, but they are application-specific and are not general-purpose software packages. However, 'sorn' is a flexible package that allows researchers to develop the network of their interest, providing them the freedom to choose the combination of plasticity rules of their choice. Moreover, it is easy to integrate 'sorn' with machine learning frameworks such as PyTorch and reinforcement learning toolkits such as OpenAI Gym. Overall, 'sorn' provides a research environment for computational neuroscientists to study self-organization, adaptation, learning, memory, and behavior of brain circuits by reverse-engineering neural plasticity mechanisms.
+There are few other open source codes [sorn v1](https://github.com/delpapa/SORN), [sorn v2](https://github.com/delpapa/SORN_V2), for SORN networks, but they are application-specific and are not general-purpose software packages. However, `sorn` is a flexible package that allows researchers to develop the network of their interest, providing them the freedom to choose the combination of plasticity rules of their choice. Moreover, it is easy to integrate `sorn` with machine learning frameworks such as PyTorch and reinforcement learning toolkits such as OpenAI Gym. Overall, `sorn` provides a research environment for computational neuroscientists to study self-organization, adaptation, learning, memory, and behavior of brain circuits by reverse-engineering neural plasticity mechanisms.
 
 
 ## Library Overview
@@ -81,7 +81,7 @@ $T_{i}(t)$ - Firing threshold of the neuron $i$ at time $t$
 
 ### Spike Timing Dependent Plasticity
 
[…]
