### 2021-09-08 #3545 @rkurchin
@Saran-nns, looks fantastic! Please just fix the couple of missing DOI's identified above and the manuscript should be pretty much set.

The next step will be to make a new tagged release of your package and archive that release (e.g. on Zenodo or figshare). Once you've done that, post the associated DOI here. Make sure the metadata matches your manuscript here.
