+++ paper/paper.bib
@@ -5,6 +5,7 @@
   volume={3},
   pages={23},
   year={2009},
+  doi = {10.3389/neuro.10.019.2009}
   publisher={Frontiers}
 }
 
@@ -33,6 +34,7 @@
   title={Special issue on echo state networks and liquid state machines.},
   author={Jaeger, Herbert and Maass, Wolfgang and Principe, Jose},
   year={2007},
+  doi = {10.1016/j.neunet.2007.04.001},
   publisher={Elsevier Science}
 }
 
