+++ paper/paper.bib
@@ -5,7 +5,7 @@
   volume={3},
   pages={23},
   year={2009},
-  doi = {10.3389/neuro.10.019.2009}
+  doi = {10.3389/neuro.10.019.2009},
   publisher={Frontiers}
 }
 
