### 2021-09-08 #3545 @Saran-nns
Hi @rkurchin,
the package is released with the tag v0.6.2-beta and the corresponding doi is DOI: 10.5281/zenodo.5496017

Thank you :)
