### 2021-09-10 #3545 @arfon
@janfreyberg, @janfb – many thanks for your reviews here and to @rkurchin for editing their first submission! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
